backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
playwright>=1.44.0
anthropic>=0.28.0
fpdf2>=2.7.9
Pillow>=10.3.0
python-dotenv>=1.0.1
